+++ crypt/input.txt
@@ -1,2 +1,2 @@
 This is secret. Message about "_" symbol!
-texttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttext
+texttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttexttext
