+++ crypt/package.json
@@ -17,11 +17,6 @@
     "stream": "0.0.2"
   },
   "devDependencies": {
-    "jest": "^27.3.1",
-    "jest": {
-      "collectCoverageFrom": [
-        "*/*.js"
-      ]
-    }
+    "jest": "^27.3.1"
   }
 }
